+++ projects/komponentkartan/polyfills.ts
@@ -1,5 +1,0 @@
-/** Smooth scrolling pollyfill */
-import smoothscroll from 'smoothscroll-polyfill';
-
-// kick off the polyfill!
-smoothscroll.polyfill();