+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed jquery
 - Refactored back-to-top from javascript to Angular component
 - Added toggle-button component 
+- Fixed a bug were elements were focused in collapsed action panels.
 
 ## 2.6.0 
 - Fixed smoother animations in list component