+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed a bug were elements were focused in collapsed action panels.
 - Bugfix for card-section animation in IE.
 - FilterText updated with new icon (themebased)
+- Added a new component Searchresult
 
 ## 2.6.0 
 - Fixed smoother animations in list component